refactor: . move QueryLayout to a separate file
